multi-currency bug fixes

+++ app/src/main/java/com/perseverance/phando/home/mediadetails/payment/PurchaseOption.kt
@@ -13,7 +13,7 @@ data class PurchaseOption(
         val note: String,
         var mediaTitle: String?,
         val payment_info: PaymentInfo,
-        val final_price : Int,
+        val final_price : Float,
         val value: Float,
         val isWallet : Boolean=false
 ) : Parcelable
